Improve some other things
Now there's a function used to print slightly indented info messages because
it looks cool and library validation is wayyyy faster using a method similar
to the one used in the asset validation.

The start subcommand doesn't work anymore but I'll work on it as soon as I
finish the download code.

Also now every message is coherent always first saying "Validating..." and
then downloading, in case an asset might be missing. This furter simplified
the script.

I really have to know how to split long strings though.